# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SignalWatch – Healthcare Social Listening API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  paused: "paused",
  archived: "archived",
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  totalSignals: number;
  safetyAlerts: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

/**
 * @nullable
 */
export type UpdateProjectBodyStatus =
  | (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus]
  | null;

export const UpdateProjectBodyStatus = {
  active: "active",
  paused: "paused",
  archived: "archived",
} as const;

export interface UpdateProjectBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  status?: UpdateProjectBodyStatus;
}

export interface SourceCount {
  sourceType: string;
  count: number;
}

/**
 * @nullable
 */
export type SignalSentiment =
  | (typeof SignalSentiment)[keyof typeof SignalSentiment]
  | null;

export const SignalSentiment = {
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
} as const;

export interface Entity {
  text: string;
  type: string;
  confidence: number;
}

export interface Signal {
  id: number;
  projectId: number;
  sourceId: number;
  sourceType: string;
  /** @nullable */
  externalId?: string | null;
  content: string;
  /** @nullable */
  author?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  sentiment?: SignalSentiment;
  /** @nullable */
  sentimentScore?: number | null;
  /** @nullable */
  confidenceScore?: number | null;
  entities: Entity[];
  hasSafetyFlag: boolean;
  /** @nullable */
  safetyReason?: string | null;
  hasPiiFlag: boolean;
  piiTypes: string[];
  isAnalyzed: boolean;
  matchedKeywords: string[];
  createdAt: string;
}

export interface ProjectSummary {
  projectId: number;
  totalSignals: number;
  safetyAlerts: number;
  piiFlags: number;
  positiveCount: number;
  negativeCount: number;
  neutralCount: number;
  sourceCounts: SourceCount[];
  recentSignals: Signal[];
}

export interface Keyword {
  id: number;
  projectId: number;
  term: string;
  createdAt: string;
}

export interface CreateKeywordBody {
  term: string;
}

export type SourceConfig = { [key: string]: unknown };

export type SourceLatency = (typeof SourceLatency)[keyof typeof SourceLatency];

export const SourceLatency = {
  realtime: "realtime",
  daily: "daily",
  weekly: "weekly",
} as const;

export interface Source {
  id: number;
  projectId: number;
  engineTypeId: number;
  engineTypeName: string;
  name: string;
  config?: SourceConfig;
  latency: SourceLatency;
  isEnabled: boolean;
  /** @nullable */
  lastRunAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateSourceBodyConfig = { [key: string]: unknown };

export type CreateSourceBodyLatency =
  (typeof CreateSourceBodyLatency)[keyof typeof CreateSourceBodyLatency];

export const CreateSourceBodyLatency = {
  realtime: "realtime",
  daily: "daily",
  weekly: "weekly",
} as const;

export interface CreateSourceBody {
  engineTypeId: number;
  name: string;
  config?: CreateSourceBodyConfig;
  latency: CreateSourceBodyLatency;
}

/**
 * @nullable
 */
export type UpdateSourceBodyConfig = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type UpdateSourceBodyLatency =
  | (typeof UpdateSourceBodyLatency)[keyof typeof UpdateSourceBodyLatency]
  | null;

export const UpdateSourceBodyLatency = {
  realtime: "realtime",
  daily: "daily",
  weekly: "weekly",
} as const;

export interface UpdateSourceBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  config?: UpdateSourceBodyConfig;
  /** @nullable */
  latency?: UpdateSourceBodyLatency;
  /** @nullable */
  isEnabled?: boolean | null;
}

export interface TriggerResult {
  signalsCollected: number;
  message: string;
}

export interface SignalListResponse {
  signals: Signal[];
  total: number;
  limit: number;
  offset: number;
}

export interface TimelinePoint {
  date: string;
  total: number;
  positive: number;
  negative: number;
  neutral: number;
  safetyAlerts: number;
}

export interface SentimentBreakdown {
  positive: number;
  negative: number;
  neutral: number;
  unanalyzed: number;
  /** @nullable */
  avgScore?: number | null;
}

export interface EntityCount {
  text: string;
  type: string;
  count: number;
}

export interface BatchAnalysisResult {
  analyzed: number;
  errors: number;
  message: string;
}

export type EngineTypeConfigSchema = { [key: string]: unknown };

export interface EngineType {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  configSchema?: EngineTypeConfigSchema;
  isBuiltIn: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateEngineTypeBodyConfigSchema = { [key: string]: unknown };

export interface CreateEngineTypeBody {
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  configSchema?: CreateEngineTypeBodyConfigSchema;
}

/**
 * @nullable
 */
export type UpdateEngineTypeBodyConfigSchema = {
  [key: string]: unknown;
} | null;

export interface UpdateEngineTypeBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  configSchema?: UpdateEngineTypeBodyConfigSchema;
}

export type ListSignalsParams = {
  /**
   * @nullable
   */
  sourceType?: string | null;
  /**
   * @nullable
   */
  sentiment?: string | null;
  /**
   * @nullable
   */
  hasSafetyFlag?: string | null;
  /**
   * @nullable
   */
  limit?: number | null;
  /**
   * @nullable
   */
  offset?: number | null;
};

export type GetSignalTimelineParams = {
  /**
   * @nullable
   */
  days?: number | null;
};

export type GetSentimentBreakdownParams = {
  /**
   * @nullable
   */
  days?: number | null;
};

export type GetTopEntitiesParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};