# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SignalWatch – Healthcare Social Listening API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all monitoring projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new monitoring project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /projects/{id}/summary:
    get:
      operationId: getProjectSummary
      tags: [projects]
      summary: Get signal summary stats for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
        "404":
          description: Not found

  # ─── Keywords ────────────────────────────────────────────────────────────────
  /projects/{projectId}/keywords:
    get:
      operationId: listKeywords
      tags: [keywords]
      summary: List keywords for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: createKeyword
      tags: [keywords]
      summary: Add a keyword to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeywordBody"
      responses:
        "201":
          description: Created keyword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
  /projects/{projectId}/keywords/{id}:
    delete:
      operationId: deleteKeyword
      tags: [keywords]
      summary: Remove a keyword from a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Sources ─────────────────────────────────────────────────────────────────
...
  /engine-types:
    get:
      operationId: listEngineTypes
      tags: [engine-types]
      summary: List all available engine/source types
      responses:
        "200":
          description: List of engine types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EngineType"
    post:
      operationId: createEngineType
      tags: [engine-types]
      summary: Register a new engine type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEngineTypeBody"
      responses:
        "201":
          description: Created engine type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngineType"
  /engine-types/{id}:
    patch:
      operationId: updateEngineType
      tags: [engine-types]
      summary: Update an engine type configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEngineTypeBody"
      responses:
        "200":
          description: Updated engine type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngineType"
    delete:
      operationId: deleteEngineType
      tags: [engine-types]
      summary: Delete an engine type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, archived]
        totalSignals:
          type: integer
        safetyAlerts:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - status
        - totalSignals
        - safetyAlerts
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [active, paused, archived, null]
    ProjectSummary:
      type: object
      properties:
        projectId:
          type: integer
        totalSignals:
          type: integer
        safetyAlerts:
          type: integer
        piiFlags:
          type: integer
        positiveCount:
          type: integer
        negativeCount:
          type: integer
        neutralCount:
          type: integer
        sourceCounts:
          type: array
          items:
            $ref: "#/components/schemas/SourceCount"
        recentSignals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
      required:
        - projectId
        - totalSignals
        - safetyAlerts
        - piiFlags
        - positiveCount
        - negativeCount
        - neutralCount
        - sourceCounts
        - recentSignals
    SourceCount:
      type: object
      properties:
        sourceType:
          type: string
        count:
          type: integer
      required:
        - sourceType
        - count

    # ─── Keyword ──────────────────────────────────────────────────────────────
    Keyword:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        term:
          type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - term
        - createdAt
    CreateKeywordBody:
      type: object
      properties:
        term:
          type: string
      required:
        - term

    # ─── Source ───────────────────────────────────────────────────────────────
...
    Signal:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        sourceId:
          type: integer
        sourceType:
          type: string
        externalId:
          type: ["string", "null"]
        content:
          type: string
        author:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        sentiment:
          type: ["string", "null"]
          enum: [positive, negative, neutral, null]
        sentimentScore:
          type: ["number", "null"]
        confidenceScore:
          type: ["number", "null"]
        entities:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        hasSafetyFlag:
          type: boolean
        safetyReason:
          type: ["string", "null"]
        hasPiiFlag:
          type: boolean
        piiTypes:
          type: array
          items:
            type: string
        isAnalyzed:
          type: boolean
        matchedKeywords:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - sourceId
        - sourceType
        - content
        - hasSafetyFlag
        - hasPiiFlag
        - isAnalyzed
        - entities
        - piiTypes
        - matchedKeywords
        - createdAt
    Entity:
      type: object
      properties:
        text:
          type: string
        type:
          type: string
        confidence:
          type: number
      required:
        - text
        - type
        - confidence
...
    EngineType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        configSchema:
          type: object
          additionalProperties: true
        isBuiltIn:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - slug
        - isBuiltIn
        - createdAt
        - updatedAt
    CreateEngineTypeBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        configSchema:
          type: object
          additionalProperties: true
      required:
        - name
        - slug
    UpdateEngineTypeBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        configSchema:
          type: ["object", "null"]
          additionalProperties: true